pyproject.toml
[tool.poetry]
name = "flat-finder"
version = "0.1.0"
description = ""
authors = ["philippbergmann <philipp.bgm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Scrapy = "^2.7.1"
requests = "^2.28.1"
undetected-chromedriver = "^3.1.7"
beautifulsoup4 = "^4.11.1"
slack-sdk = "^3.19.5"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

